refactor reusable components to separate files
