+++ frontend/system/src/Admin/DrugEditForm/AddFieldButton.tsx
@@ -1,0 +1,26 @@
+import { FlatButton } from '@common/components';
+import { PlusCircleIcon } from '@common/icons';
+import React from 'react';
+
+export const AddFieldButton = ({
+  label,
+  onClick,
+}: {
+  label: string;
+  onClick: () => void;
+}) => {
+  return (
+    <FlatButton
+      startIcon={<PlusCircleIcon />}
+      label={label}
+      onClick={onClick}
+      disableFocusRipple
+      sx={{
+        marginLeft: '20px',
+        '&.Mui-focusVisible': {
+          backgroundColor: '#e95c3029',
+        },
+      }}
+    />
+  );
+};
