+++ frontend/system/src/Admin/DrugEditForm/CategoryDropdown.tsx
@@ -1,0 +1,31 @@
+import React from 'react';
+import { MenuItem, Option, Select } from '@common/ui';
+
+export const CategoryDropdown = ({
+  value,
+  options,
+  onChange,
+  getOptionDisabled,
+}: {
+  value: string;
+  options: Option[];
+  onChange: (value: string) => void;
+  getOptionDisabled: (o: Option) => boolean;
+}) => (
+  <Select
+    autoFocus
+    value={value}
+    onChange={e => onChange(e.target.value)}
+    options={options}
+    renderOption={(option: Option) => (
+      <MenuItem
+        key={option.value}
+        value={option.value}
+        disabled={getOptionDisabled(option)}
+      >
+        {option.label}
+      </MenuItem>
+    )}
+    sx={{ width: '100%' }}
+  />
+);
