+++ frontend/system/src/Admin/DrugEditForm/index.ts
@@ -1,0 +1,1 @@
+export * from './DrugEditForm';
